# Conflicts:
#	src/main/java/ch/uzh/ifi/seal/soprafs20/constant/GameStatus.java
#	src/main/java/ch/uzh/ifi/seal/soprafs20/controller/RoundController.java
#	src/main/java/ch/uzh/ifi/seal/soprafs20/service/HintValidationService.java
#	src/main/java/ch/uzh/ifi/seal/soprafs20/service/RoundService.java

+++ src/main/java/ch/uzh/ifi/seal/soprafs20/rest/dto/player/PlayerPutDTO.java
@@ -9,7 +9,7 @@ public class PlayerPutDTO {
     private PlayerStatus status;
     private int score;
     private PlayerRole role;
-    private int elapsedTime;
+    private long elapsedTime;
 
     public String getName() {
         return name;
@@ -43,11 +43,11 @@ public class PlayerPutDTO {
         this.role = role;
     }
 
-    public int getElapsedTime() {
+    public long getElapsedTime() {
         return elapsedTime;
     }
 
-    public void setElapsedTime(int elapsedTime) {
+    public void setElapsedTime(long elapsedTime) {
         this.elapsedTime = elapsedTime;
     }
 }
